+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fix `index` to not generate empty module
 * Add dots to letter answers in `data-science`
+* Add `BakeEmbeddedExerciseQuestion` for baking exercise-like things
+* Bake ordered lists as embedded exercise questions in `computer-science`
 
 ## [v2.23.0] - 2024-09-09